Moved kind to layers
